+++ src/Facades/X.php
@@ -7,16 +7,16 @@ use Illuminate\Support\Facades\Facade;
 /**
  * Class X
  *
- * Facade for TwitterService (X).
+ * Facade for XService (X).
  *
  * @method static string getAuthorizationUrl(string $redirectUri, array $scopes = [], ?string $state = null)
  * @method static array handleCallback(string $code, string $redirectUri, string $codeVerifier)
- * @method static \Mantix\LaravelSocialMediaPublisher\Services\TwitterService forConnection(\Mantix\LaravelSocialMediaPublisher\Models\SocialMediaConnection $connection)
+ * @method static \Mantix\LaravelSocialMediaPublisher\Services\XService forConnection(\Mantix\LaravelSocialMediaPublisher\Models\SocialMediaConnection $connection)
  *
- * @see \Mantix\LaravelSocialMediaPublisher\Services\TwitterService
+ * @see \Mantix\LaravelSocialMediaPublisher\Services\XService
  */
 class X extends Facade {
     protected static function getFacadeAccessor() {
-        return \Mantix\LaravelSocialMediaPublisher\Services\TwitterService::class;
+        return \Mantix\LaravelSocialMediaPublisher\Services\XService::class;
     }
 }